backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.0.3

# AI and ML
google-generativeai==0.3.1
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# Pydantic Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1

# Note: httpx version must be <0.25.0 for supabase compatibility